Client: Propagate curLatLng state through so we can display it on Sidebar/Form
