+++ client/src/Sidebar/Feed/Feed.js
@@ -1,5 +1,5 @@
 import React from 'react';
-export default class Feed extends React.Component{
+export default class Feed extends React.Component {
     constructor(props) {
         super(props)
     }
@@ -7,7 +7,7 @@ export default class Feed extends React.Component{
         return (
             <div>
                 <ul>
-                    {this.props.elements.map(x => (<li>{x}</li>))}
+                    {this.props.elements.map(x => (<li key={x}>{x}</li>))}
                 </ul>
             </div>
         )
